moved profile_models and phase_space_models into subpackage
