+++ halotools/empirical_models/__init__.py
@@ -16,6 +16,6 @@ from .assembias_models import *
 from .hod_designer import *
 # from .conc_mass_models import *
 # from .monte_carlo_phase_space import *
-from .phase_space_models import *
+# from .phase_space_models import *
 # from .profile_helpers import *
-from .profile_models import *
+# from .profile_models import *
