+++ halotools/empirical_models/phase_space_marf/phase_space_models.py
@@ -14,11 +14,12 @@ import numpy as np
 from astropy.table import Table
 
 from .profile_models import *
-from .phase_space_marf.velocity_models import *
-from .phase_space_marf.monte_carlo_phase_space import *
-from . import model_defaults
+from .velocity_models import *
+from .monte_carlo_phase_space import *
 
-from ..sim_manager import sim_defaults
+from .. import model_defaults
+
+from ...sim_manager import sim_defaults
 
 
 class NFWPhaseSpace(NFWProfile, NFWJeansVelocity, MonteCarloGalProf):
