+++ halotools/empirical_models/phase_space_marf/profile_models.py
@@ -16,13 +16,13 @@ from astropy import units as u
 from astropy.constants import G
 newtonG = G.to(u.km*u.km*u.Mpc/(u.Msun*u.s*u.s))
 
-from . import model_defaults
-from .phase_space_marf.conc_mass_models import ConcMass
-from .phase_space_marf.profile_helpers import *
+from .. import model_defaults
+from .conc_mass_models import ConcMass
+from .profile_helpers import *
 
-from ..utils.array_utils import convert_to_ndarray
-from ..custom_exceptions import *
-from ..sim_manager import sim_defaults 
+from ...utils.array_utils import convert_to_ndarray
+from ...custom_exceptions import *
+from ...sim_manager import sim_defaults 
 
 
 __author__ = ['Andrew Hearin', 'Benedikt Diemer']
